requirements.txt
flask==2.3.3
Pillow==10.0.1
ImageHash==4.3.1
praw==7.7.1
requests==2.31.0
numpy==2.2.6
python-dotenv==1.0.0
pytest==7.4.0
gunicorn==21.2.0